+++ client/state/products-list/actions.js
@@ -6,7 +6,6 @@ import { mapValues } from 'lodash';
 /**
  * Internal dependencies
  */
-
 import wpcom from 'lib/wp';
 import { createProductObject } from './assembler';
 import {
@@ -14,6 +13,8 @@ import {
 	PRODUCTS_LIST_REQUEST,
 	PRODUCTS_LIST_REQUEST_FAILURE,
 } from 'state/action-types';
+
+import 'state/products-list/init';
 
 export function receiveProductsList( productsList ) {
 	return {
