+++ client/state/products-list/selectors/get-products-list.js
@@ -1,0 +1,8 @@
+/**
+ * Internal dependencies
+ */
+import 'state/products-list/init';
+
+export function getProductsList( state ) {
+	return state.productsList.items;
+}
