+++ client/state/products-list/selectors/is-products-list-fetching.js
@@ -1,0 +1,8 @@
+/**
+ * Internal dependencies
+ */
+import 'state/products-list/init';
+
+export function isProductsListFetching( state ) {
+	return state.productsList.isFetching;
+}
